+++ apps/vmq_server/src/vmq_mqtt_fsm.erl
@@ -21,7 +21,8 @@
     init/3,
     data_in/2,
     msg_in/2,
-    info/2
+    info/2,
+    subscriber/1
 ]).
 
 -define(IS_PROTO_4(X), X =:= 4; X =:= 132).
@@ -91,6 +92,11 @@
 -export_type([state/0]).
 -define(state_val(Key, Args, State), prop_val(Key, Args, State#state.Key)).
 -define(cap_val(Key, Args, State), prop_val(Key, Args, CAPSettings#cap_settings.Key)).
+
+subscriber({connected, State}) ->
+    State#state.subscriber_id;
+subscriber(_) ->
+    undefined.
 
 init(
     Peer,